requirements.txt
openai
tk
rich
pywebview